+++ frontend/src/api-types.ts
@@ -417,6 +417,23 @@ export interface paths {
         patch?: never;
         trace?: never;
     };
+    "/repositories/{repo_id}/integration": {
+        parameters: {
+            query?: never;
+            header?: never;
+            path?: never;
+            cookie?: never;
+        };
+        get?: never;
+        put?: never;
+        /** Link Repo Integration */
+        post: operations["link_repo_integration_repositories__repo_id__integration_post"];
+        delete?: never;
+        options?: never;
+        head?: never;
+        patch?: never;
+        trace?: never;
+    };
     "/claims/{claim_id}": {
         parameters: {
             query?: never;
@@ -1638,6 +1655,11 @@ export interface components {
             /** Layer */
             layer: string;
         };
+        /** RepoLink */
+        RepoLink: {
+            /** Integration Id */
+            integration_id?: string | null;
+        };
         /** RequestApproval */
         RequestApproval: {
             /** To */
@@ -2639,6 +2661,43 @@ export interface operations {
             cookie?: never;
         };
         requestBody?: never;
+        responses: {
+            /** @description Successful Response */
+            200: {
+                headers: {
+                    [name: string]: unknown;
+                };
+                content: {
+                    "application/json": unknown;
+                };
+            };
+            /** @description Validation Error */
+            422: {
+                headers: {
+                    [name: string]: unknown;
+                };
+                content: {
+                    "application/json": components["schemas"]["HTTPValidationError"];
+                };
+            };
+        };
+    };
+    link_repo_integration_repositories__repo_id__integration_post: {
+        parameters: {
+            query?: never;
+            header?: {
+                authorization?: string | null;
+            };
+            path: {
+                repo_id: string;
+            };
+            cookie?: never;
+        };
+        requestBody: {
+            content: {
+                "application/json": components["schemas"]["RepoLink"];
+            };
+        };
         responses: {
             /** @description Successful Response */
             200: {